Implement RDMA READ limit negotiation

+++ src/kmod/urdma.h
@@ -49,7 +49,6 @@
 
 #include "urdma_kabi.h"
 #include "backports.h"
-#include "iwarp.h"
 
 #define SIW_MAX_QP		(1024 * 100)
 #define SIW_MAX_ORD		128
@@ -68,8 +67,8 @@ struct siw_devinfo {
 	u32			vendor_part_id;
 	u32			sw_version;
 	int			max_qp;
-	int			max_ord; /* max. outbound read queue depth */
-	int			max_ird; /* max. inbound read queue depth */
+	u16			max_ord; /* max. outbound read queue depth */
+	u16			max_ird; /* max. inbound read queue depth */
 
 	enum ib_device_cap_flags	cap_flags;
 	int			max_cq;
@@ -179,8 +178,8 @@ struct siw_sk_upcalls {
 
 struct siw_qp_attrs {
 	enum siw_qp_state	state;
-	u32			orq_size;
-	u32			irq_size;
+	u16			orq_size;
+	u16			irq_size;
 	u16			urdma_devid;
 	u16			urdma_qp_id;
 	u16			urdma_rxq;
